+++ CHANGES.rst
@@ -16,6 +16,7 @@ Changelog for Abilian Core
 - Fixed setupwizard.
 - added config value: BABEL_ACCEPT_LANGUAGES, to limit supported languages and
   change order during negociation
+- Switched CSS to LESS.
 
 0.1.3 (2014-02-03)
 ------------------